+++ CHANGELOG.md
@@ -17,8 +17,17 @@
     Click to see more.
   </summary>
 
+## :house: Internal
+
+- `orchestrator`
+  - [#4300](https://github.com/wix/ricos/pull/4300) Event Source API
 
 </details>
+
+## 8.72.17 (Aug 01, 2022)
+
+- `image-plugin`
+  - [#4307](https://github.com/wix/ricos/pull/4307) wix-image cleanup console log
 
 ## 8.72.16 (Aug 01, 2022)